- Changing packages to move membership requests to the team domain
