+++ nucleus/src/main/java/za/co/absa/subatomic/adapter/team/rest/MembershipRequestResource.java
@@ -1,8 +1,8 @@
-package za.co.absa.subatomic.adapter.membership_request.rest;
+package za.co.absa.subatomic.adapter.team.rest;
 
 import lombok.Data;
 import org.springframework.hateoas.ResourceSupport;
-import za.co.absa.subatomic.domain.membership_request.MembershipRequestStatus;
+import za.co.absa.subatomic.domain.team.MembershipRequestStatus;
 
 @Data
 public class MembershipRequestResource extends ResourceSupport {
