+++ nucleus/src/main/java/za/co/absa/subatomic/adapter/team/rest/NewMembershipRequestResource.java
@@ -1,4 +1,4 @@
-package za.co.absa.subatomic.adapter.membership_request.rest;
+package za.co.absa.subatomic.adapter.team.rest;
 
 import lombok.Data;
 import org.springframework.hateoas.ResourceSupport;
